중간 병합
