+++ back/src/main/java/com/onestep/back/service/goal/GoalDtlService.java
@@ -1,7 +1,7 @@
 package com.onestep.back.service.goal;
 
-import com.onestep.back.dto.GoalDTO;
-import com.onestep.back.dto.goal.CertDTO;
+import com.onestep.back.dto.goal.GoalDTO;
+import com.onestep.back.dto.upload.CertificationsDTO;
 import com.onestep.back.dto.goal.GoalDtlDTO;
 
 import java.util.List;
@@ -10,7 +10,7 @@ public interface GoalDtlService {
 
     GoalDtlDTO getGoalInfo(Long goalId);
 
-    List<CertDTO> getRecentCert(Long goalId);
+    List<CertificationsDTO> getRecentCert(Long goalId);
 
     void joinGoal(GoalDTO goalDTO);
