+++ front/src/App.jsx
@@ -6,17 +6,22 @@ import Layout from './Layout';
 import GoalList from './components/goal/GoalList';
 import GoalRegister from './components/goal/GoalRegister';
 import GoalDtl from './components/goal/GoalDtl';
+import ChatPage from './components/chat/ChatPage';
+import ChatRoom from './components/chat/ChatRoom';
 
 function App() {
   return (
     <Routes>
       <Route element={<Layout />}>
-        <Route path="/:goalid" element={<GoalDtl />} />
         <Route path="/test" element={<Test />} />
 
-        {/* 목표 관련 라우트 추가 */}
         <Route path="/" element={<GoalList />} />
         <Route path="/goals/register" element={<GoalRegister />} />
+        <Route path="/:goalid" element={<GoalDtl />} />
+
+        <Route path="/chat" element={<ChatPage />}>
+          <Route path=":chatId" element={<ChatRoom />} />
+        </Route>
       </Route>
     </Routes>
   );
